Change default insecure password to: raspberrypiqemu
Note: The default password was triggering a lockout when SSHd was enabled.

The default 'raspberrypi' password appears to trigger some protections built
into the Raspbian image, which were designed to force users to change the
password after first boot.

Since this is an ephemeral development VM, we don't really care about security,
nor do we want to lockout the 'pi' user from logging in via the QEMU getty
window.